+++ manufacturing-ai-suite/industrial-edge-insights-multimodal/.env
@@ -192,10 +192,6 @@ NGINX_CONFIG_PATH=./configs/nginx/nginx.conf
 # =========
 # Agentic config
 # =========
-# API key used by agent/storage mutating endpoints (X-API-Key header)
-# Change this value before exposing the stack beyond localhost/trusted networks.
-APM_API_KEY=change-me-apm-key
-
 # ── LLM / VLM ────────────────────────────────────────────────────────────────
 # Set LLM_MODE=fallback to run without an LLM (rule-based operation)
 LLM_MODE=llm
