+++ manufacturing-ai-suite/industrial-edge-insights-multimodal/configs/influxdb/config/influxdb.conf
@@ -115,7 +115,7 @@
   # The maximum number of tag values per tag that are allowed before writes are dropped.  This limit
   # can prevent high cardinality tag values from being written to a measurement.  This limit can be
   # disabled by setting it to 0.
-  # max-values-per-tag = 100000
+  max-values-per-tag = 1000000
 
 ###
 ### [coordinator]
