+++ manufacturing-ai-suite/industrial-edge-insights-multimodal/docker-compose-agentic.yml
@@ -177,11 +177,9 @@ services:
       http_proxy: ${http_proxy:-}
       https_proxy: ${https_proxy:-}
       no_proxy: ${no_proxy:-},apm-agent,apm-detection,ia-fusion-analytics,localhost,ia-mqtt-broker
-      APM_API_KEY: ${APM_API_KEY:-change-me-apm-key}
       AGENT_SERVICE_URL: http://apm-agent:5002
       DETECTION_SERVICE_URL: http://apm-detection:5004
       STORAGE_SERVICE_URL: http://ia-fusion-analytics:8080
-      USE_CASE_ID: ${USE_CASE}
       HOST: 0.0.0.0
       PORT: 5003
       LOG_LEVEL: ${LOG_LEVEL:-INFO}
