+++ manufacturing-ai-suite/industrial-edge-insights-multimodal/fusion-analytics/Dockerfile
@@ -18,7 +18,7 @@ RUN pip install --upgrade pip
 RUN pip install --no-cache-dir -r requirements.txt
 
 # Copy application code
-COPY fusion.py .
+COPY fusion.py api.py .
 
 # Set environment variables
 ENV PYTHONUNBUFFERED=1
